refactor(modules): Reorganized module-related models and fixed import errors
New models structure looks like this:
models/
├─ dynamic/
│  ├─ dynamic.py
│  ├─ module.py  <-- imports services
│  ├─ module_settings.py
│  ├─ module_settings_arcade.py
├─ static/
│  ├─ static.py
│  ├─ module_handlers.py  <-- imports services
