+++ mln/admin/__init__.py
@@ -9,9 +9,10 @@ from django.contrib.admin.helpers import ActionForm
 from django.db.models import Q
 
 from ..models.dynamic import Attachment, Friendship, Message, Profile, InventoryStack
-from ..models.module import Module, ModuleExecutionCost, ModuleGuestYield, ModuleMessage, ModuleOwnerYield, module_settings_classes, ModuleSaveConcertArcade, ModuleSaveSoundtrack
-from ..models.module_settings_arcade import DeliveryArcadeTile
+from ..models.dynamic.module import Module, module_settings_classes, ModuleSaveConcertArcade, ModuleSaveSoundtrack
+from ..models.dynamic.module_settings_arcade import DeliveryArcadeTile
 from ..models.static import Answer, BlueprintInfo, BlueprintRequirement, ItemInfo, ItemType, MessageBody, MessageBodyType, MessageTemplate, MessageTemplateAttachment, ModuleEditorType, ModuleHarvestYield, ModuleInfo, ModuleSetupCost, NetworkerFriendshipCondition, NetworkerFriendshipConditionSource, NetworkerMessageTriggerLegacy, NetworkerMessageAttachmentLegacy, NetworkerReply, StartingStack, Question
+from ..models.static.module_handlers import ModuleExecutionCost, ModuleGuestYield, ModuleMessage, ModuleOwnerYield
 from .make_inline import custom, inlines, make_inline
 
 # Normal but customized admin interfaces
