+++ mln/models/__init__.py
@@ -1,0 +1,5 @@
+from .dynamic import module
+from .dynamic import module_settings
+from .dynamic import module_settings_arcade
+
+from .static import module_handlers
