+++ mln/models/dynamic/__init__.py
@@ -1,0 +1,5 @@
+from .dynamic import *
+
+from . import module
+from . import module_settings
+from . import module_settings_arcade
