+++ mln/models/static/__init__.py
@@ -1,0 +1,1 @@
+from .static import *
